Struct zng_ext_config::settings::Setting

pub struct Setting { /* private fields */ }

Setting entry.

Implementations

impl Setting

pub fn key(&self) -> &ConfigKey

The config edited by this setting.

pub fn order(&self) -> u16

Position of the setting in a list of settings.

Lower numbers are listed first, two settings with the same order are sorted by display name.

pub fn name(&self) -> &BoxedVar<Txt>

Display name.

pub fn description(&self) -> &BoxedVar<Txt>

Short help text.

pub fn category(&self) -> &CategoryId

Settings category.

pub fn meta(&self) -> StateMapRef<'_, Setting>

Custom setting metadata.

pub fn value_is_set(&self) -> bool

If the value is set to an actual config variable.

Setting builders can not set the value, this can be used to indicate that the setting must be edited directly on the config file.

pub fn value(&self) -> &BoxedAnyVar

Config value.

pub fn value_type(&self) -> TypeId

Config value type.

pub fn value_downcast<T: ConfigValue>(&self) -> Option<BoxedVar<T>>

Config value, strongly typed.

pub fn can_reset(&self) -> BoxedVar<bool>

Gets a variable that indicates the current setting value is not the default.

pub fn reset(&self)

Reset the setting value.

pub fn search_index(&self, search: &str) -> Option<usize>

Gets if the setting should be included in the search and how likely it is to be an exact match (0 is exact).

If search starts with @key: matches key case sensitive, otherwise matches name or description in lower case. Note that non-key search is expected to already be lowercase.