Struct zng_ext_config::FallbackConfig

pub struct FallbackConfig<S: Config, F: Config>(/* private fields */);

Represents a copy-on-write config source that wraps two other sources, a read-write config and a read-only fallback config.

The config variables are connected to both sources, if the read-write config is not set the var will update with the fallback config, if it is set it will sync with the read-write config.

The FallbackConfig type is an Arc internally, so you can keep a cloned reference to it after moving it into CONFIG or another combinator config.

Implementations

impl<S: Config, F: Config> FallbackConfig<S, F>

pub fn new(config: S, fallback: F) -> Self

New from the read-write config and read-only fallback.

pub fn reset(&self, key: &ConfigKey)

Removes the key from the config and updates all active config variables back to the fallback value. Note that if you assign the config variable the key will be re-inserted on the config.

pub fn can_reset(&self, key: ConfigKey) -> BoxedVar<bool>

Returns a read-only var that is true when the key has an entry in the read-write config.