zng_ext_config

Trait ConfigValue

pub trait ConfigValue:
    VarValue
    + Serialize
    + DeserializeOwned { }

Marker trait for types that can stored in a Config.

This trait is already implemented for types it applies.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.