Crate zng_ext_config

Config service and sources.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 3 feature flags, 0 enabled by default.

"yaml"

Enable YAML support.

"toml"

Enable TOML support.

"ron"

Enable RON support.

Modules

settings

Settings are the config the user can directly edit, this module implements a basic settings data model.

Structs

CONFIG

Represents the app main config.

ConfigManager

Application extension that provides mouse events and service.

ConfigVars

Map of configs already bound to a variable.

FallbackConfig

Represents a copy-on-write config source that wraps two other sources, a read-write config and a read-only fallback config.

MemoryConfig

Memory only config.

RawConfigValue

Represents any entry type in a config.

ReadOnlyConfig

Config wrapper that only provides read-only variables from the inner config.

SwapConfig

Represents a config source that can swap its backing config source without disconnecting any bound keys.

SwitchConfig

Represents multiple config sources that are matched by key.

SyncConfig

Config source that auto syncs with file.

Enums

ConfigStatus

Represents the current IO status of the config.

RonValueRawError

Error converting ron::Value, RawConfigValue.

TomlValueRawError

Error converting toml::Value, RawConfigValue.

YamlValueRawError

Error converting serde_yaml::Value, RawConfigValue.

Traits

AnyConfig

Represents one or more config sources behind a dynamic reference.

Config

Represents one or more config sources.

ConfigMap

Represents a full config map in memory.

ConfigValue

Marker trait for types that can stored in a Config.

FallbackConfigReset

Reset controls of a FallbackConfig.

Type Aliases

ConfigKey

Unique key to a config entry.

JsonConfig

Represents a config source that synchronizes with a JSON file.

RonConfig

Represents a config source that synchronizes with a RON file.

TomlConfig

Represents a config source that synchronizes with a TOML file.

YamlConfig

Represents a config source that synchronizes with a YAML file.