Module zng_ext_config::settings

Settings are the config the user can directly edit, this module implements a basic settings data model.

The settings editor widget is not implemented here, this module bridges config implementers with settings UI implementers.

Structs

CategoriesBuilder

Setting categories builder.

Category

Settings category.

CategoryBuilder

Category entry builder.

CategoryId

Unique ID of a Category.

SETTINGS

Settings metadata service.

Setting

Setting entry.

SettingBuilder

Setting entry builder.

SettingsBuilder

Settings builder.